pyproject.toml
[build-system]
requires = ['setuptools', 'setuptools-scm']
build-backend = 'setuptools.build_meta'

[project]
name = 'concatmap'
version = '2.0.0'
authors = [
    {name = 'Daryl Gohl'},
    {name = 'Paul W. Egeler, M.S.'}
]
description = 'A program to map and plot reads against a circular reference.'
license = 'MIT'
requires-python = '>=3.10'
dynamic = ['readme']
dependencies = [
    'biopython>=1.79',
    'matplotlib>=3.5.1',
    'scipy>=1.7.3',
    'numpy>=1.21.5',
    'pysam>=0.19.1'
]

[external]
dependencies = [
    'dep:generic/minimap2'
]

[project.urls]
Homepage = 'https://github.com/darylgohl/ConcatMap'
Issues = 'https://github.com/darylgohl/ConcatMap/issues'

[project.scripts]
concatmap = 'concatmap.cli:main'

[tool.setuptools.dynamic]
readme = {file = ['README.md'], content-type = 'text/markdown'}

[tool.isort]
force_single_line = true

[tool.yapf]
based_on_style = 'pep8'
